Move limits to its own module
