+++ eel/examples/eel-node/cli.rs
@@ -1,7 +1,7 @@
 use bitcoin::secp256k1::PublicKey;
 use chrono::{DateTime, Utc};
 use colored::Colorize;
-use eel::payment::LiquidityLimit;
+use eel::limits::LiquidityLimit;
 use rustyline::config::Builder;
 use rustyline::error::ReadlineError;
 use rustyline::history::DefaultHistory;
