+++ eel/src/lib.rs
@@ -8,6 +8,7 @@ pub mod errors;
 pub mod interfaces;
 pub mod key_derivation;
 pub mod keys_manager;
+pub mod limits;
 pub mod lsp;
 mod migrations;
 pub mod node_info;
@@ -44,10 +45,11 @@ use crate::interfaces::{EventHandler, ExchangeRate, ExchangeRateProvider, Remote
 pub use crate::invoice::InvoiceDetails;
 use crate::invoice::{create_invoice, validate_invoice, CreateInvoiceParams};
 use crate::keys_manager::init_keys_manager;
+use crate::limits::PaymentAmountLimits;
 use crate::logger::LightningLogger;
 use crate::lsp::{calculate_fee, LspClient, LspFee};
 use crate::node_info::{estimate_max_incoming_payment_size, get_channels_info, NodeInfo};
-use crate::payment::{FiatValues, Payment, PaymentAmountLimits, PaymentState, PaymentType};
+use crate::payment::{FiatValues, Payment, PaymentState, PaymentType};
 use crate::random::generate_random_bytes;
 use crate::rapid_sync_client::RapidSyncClient;
 use crate::storage_persister::StoragePersister;
